+++ lib/rules/no-top-level-state.ts
@@ -1,14 +1,69 @@
 // SPDX-License-Identifier: ISC
 
 import type {Rule} from 'eslint';
+import type {Expression, VariableDeclarator} from 'estree';
 
 import {isTopLevel} from '../helpers';
 
+type Options = {
+  readonly allow: ReadonlyArray<string>;
+};
+
+const allowOption = {
+  enum: ['ArrayExpression', 'ObjectExpression'],
+  default: [],
+  always: [
+    'ArrowFunctionExpression',
+    'AssignmentExpression',
+    'AwaitExpression',
+    'BinaryExpression',
+    'CallExpression',
+    'ChainExpression',
+    'ConditionalExpression',
+    'FunctionExpression',
+    'Identifier',
+    'ImportExpression',
+    'Literal',
+    'LogicalExpression',
+    'MemberExpression',
+    'SequenceExpression',
+    'TaggedTemplateExpression',
+    'TemplateLiteral',
+    'ThisExpression',
+    'UnaryExpression',
+    'UpdateExpression'
+  ]
+};
+
+const disallowedArray = {
+  id: '1',
+  message: 'Arrays are stateful (mutable) and not allowed at the top level'
+};
+const disallowedObject = {
+  id: '2',
+  message: 'Objects are stateful (mutable) and not allowed at the top level'
+};
 const disallowedRegexp = {
   id: '0',
   message:
     'Regular expressions with the `g` or `y` flag are stateful and not allowed at the top level'
 };
+const disallowedVar = {
+  id: '3',
+  message:
+    "Variables declared with 'var' are stateful and not allowed at the top level"
+};
+const disallowedLet = {
+  id: '4',
+  message:
+    "Variables declared with 'let' are stateful and not allowed at the top level"
+};
+
+function isInitialized(
+  node: VariableDeclarator
+): node is VariableDeclarator & {init: Expression} {
+  return node.init !== null;
+}
 
 export const noTopLevelState: Rule.RuleModule = {
   meta: {
@@ -18,12 +73,40 @@ export const noTopLevelState: Rule.RuleModule = {
       recommended: true,
       url: 'https://github.com/ericcornelissen/eslint-plugin-top/blob/main/docs/rules/no-top-level-state.md'
     },
-    schema: [],
+    schema: [
+      {
+        type: 'object',
+        properties: {
+          allow: {
+            description:
+              'Configure what is allowed to be assigned to variables',
+            type: 'array',
+            minItems: 0,
+            items: {
+              enum: allowOption.enum
+            }
+          }
+        }
+      }
+    ],
     messages: {
-      [disallowedRegexp.id]: disallowedRegexp.message
+      [disallowedArray.id]: disallowedArray.message,
+      [disallowedObject.id]: disallowedObject.message,
+      [disallowedRegexp.id]: disallowedRegexp.message,
+      [disallowedVar.id]: disallowedVar.message,
+      [disallowedLet.id]: disallowedLet.message
     }
   },
   create: (context) => {
+    const [provided] = context.options as Partial<Options>[]; // type-coverage:ignore-line
+
+    const options: Options = {
+      allow: [
+        ...allowOption.always,
+        ...(provided?.allow || allowOption.default)
+      ]
+    };
+
     return {
       Literal: (node) => {
         if (!('regex' in node)) {
@@ -45,6 +128,50 @@ export const noTopLevelState: Rule.RuleModule = {
           node,
           messageId: disallowedRegexp.id
         });
+      },
+      VariableDeclaration: (node) => {
+        if (!isTopLevel(node)) {
+          return;
+        }
+
+        switch (node.kind) {
+          case 'var': {
+            context.report({node, messageId: disallowedVar.id});
+            break;
+          }
+          case 'let': {
+            context.report({node, messageId: disallowedLet.id});
+            break;
+          }
+          case 'const':
+          case 'using':
+          case 'await using':
+        }
+      },
+      VariableDeclarator: (node) => {
+        if (!isInitialized(node)) {
+          return;
+        }
+
+        if (options.allow.includes(node.init.type)) {
+          return;
+        }
+
+        if (!isTopLevel(node)) {
+          return;
+        }
+
+        if (node.init.type === 'ArrayExpression') {
+          context.report({
+            node: node.init,
+            messageId: disallowedArray.id
+          });
+        } else {
+          context.report({
+            node: node.init,
+            messageId: disallowedObject.id
+          });
+        }
       }
     };
   }