+++ tests/unit/no-top-level-state.test.ts
@@ -5,9 +5,45 @@ import {RuleTester} from 'eslint';
 import {trimTestCases} from './helpers';
 import {noTopLevelState} from '../../lib/rules/no-top-level-state';
 
+const options: {
+  [key: string]: {
+    allow?: string[];
+  };
+} = {
+  allowArray: {
+    allow: ['ArrayExpression']
+  },
+  allowObject: {
+    allow: ['ObjectExpression']
+  }
+};
+
 const valid: RuleTester.ValidTestCase[] = [
   // Not top level
   ...[
+    {
+      code: `
+        function f() {
+          const boolean = true;
+          const number = 3.14;
+          const string = 'Hello world!';
+        }
+      `
+    },
+    {
+      code: `
+        function f() {
+          const array = [];
+        }
+      `
+    },
+    {
+      code: `
+        function f() {
+          const object = [];
+        }
+      `
+    },
     {
       code: `
         function f() {
@@ -70,6 +106,22 @@ const valid: RuleTester.ValidTestCase[] = [
           const regexpUnicodeSets = /bar/v;
         }
       `
+    },
+    {
+      code: `
+        function f() {
+          var uninitialized;
+          var initialized1 = 'foobar';
+        }
+      `
+    },
+    {
+      code: `
+        function f() {
+          let uninitialized;
+          let initialized1 = 'foobar';
+        }
+      `
     }
   ],
 
@@ -101,6 +153,38 @@ const valid: RuleTester.ValidTestCase[] = [
     }
   ],
 
+  // Top-level array
+  ...[
+    {
+      code: `
+        const array = [/*empty*/];
+      `,
+      options: [options.allowArray]
+    },
+    {
+      code: `
+        const array = ["non", "empty"];
+      `,
+      options: [options.allowArray]
+    }
+  ],
+
+  // Top-level object
+  ...[
+    {
+      code: `
+        const object = {/*empty*/};
+      `,
+      options: [options.allowObject]
+    },
+    {
+      code: `
+        const object = { non: "empty" };
+      `,
+      options: [options.allowObject]
+    }
+  ],
+
   // Top-level stateless regular expressions
   ...[
     {
@@ -128,6 +212,100 @@ const valid: RuleTester.ValidTestCase[] = [
 ];
 
 const invalid: RuleTester.InvalidTestCase[] = [
+  // Top-level array
+  ...[
+    {
+      code: `
+        const array = [/*empty*/];
+      `,
+      errors: [
+        {
+          messageId: '1',
+          line: 1,
+          column: 15,
+          endLine: 1,
+          endColumn: 26
+        }
+      ]
+    },
+    {
+      code: `
+        const array = ["non", "empty"];
+      `,
+      errors: [
+        {
+          messageId: '1',
+          line: 1,
+          column: 15,
+          endLine: 1,
+          endColumn: 31
+        }
+      ]
+    },
+    {
+      code: `
+        const array = ["even", "when", "objects", "are", "allowed"];
+      `,
+      options: [options.allowObject],
+      errors: [
+        {
+          messageId: '1',
+          line: 1,
+          column: 15,
+          endLine: 1,
+          endColumn: 60
+        }
+      ]
+    }
+  ],
+
+  // Top-level object
+  ...[
+    {
+      code: `
+        const object = {/*empty*/};
+      `,
+      errors: [
+        {
+          messageId: '2',
+          line: 1,
+          column: 16,
+          endLine: 1,
+          endColumn: 27
+        }
+      ]
+    },
+    {
+      code: `
+        const object = { non: "empty" };
+      `,
+      errors: [
+        {
+          messageId: '2',
+          line: 1,
+          column: 16,
+          endLine: 1,
+          endColumn: 32
+        }
+      ]
+    },
+    {
+      code: `
+        const object = { even: "when arrays are allowed" };
+      `,
+      options: [options.allowArray],
+      errors: [
+        {
+          messageId: '2',
+          line: 1,
+          column: 16,
+          endLine: 1,
+          endColumn: 51
+        }
+      ]
+    }
+  ],
+
   // Top-level stateful regular expressions
   ...[
     {
@@ -155,6 +333,66 @@ const invalid: RuleTester.InvalidTestCase[] = [
           column: 22,
           endLine: 1,
           endColumn: 31
+        }
+      ]
+    }
+  ],
+
+  // Top-level variable
+  ...[
+    {
+      code: `
+        var uninitialized;
+      `,
+      errors: [
+        {
+          messageId: '3',
+          line: 1,
+          column: 1,
+          endLine: 1,
+          endColumn: 19
+        }
+      ]
+    },
+    {
+      code: `
+        var initialized = 'foobar';
+      `,
+      errors: [
+        {
+          messageId: '3',
+          line: 1,
+          column: 1,
+          endLine: 1,
+          endColumn: 28
+        }
+      ]
+    },
+    {
+      code: `
+        let uninitialized;
+      `,
+      errors: [
+        {
+          messageId: '4',
+          line: 1,
+          column: 1,
+          endLine: 1,
+          endColumn: 19
+        }
+      ]
+    },
+    {
+      code: `
+        let initialized = 'foobar';
+      `,
+      errors: [
+        {
+          messageId: '4',
+          line: 1,
+          column: 1,
+          endLine: 1,
+          endColumn: 28
         }
       ]
     }